% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{do_the_thing}
\alias{do_the_thing}
\title{solves the exercise given the path to a valid sql database containing the annotated genomic data.}
\usage{
do_the_thing(path)
}
\description{
solves the exercise given the path to a valid sql database containing the annotated genomic data.
}
